Add mocha test framework.
I've observed a bug where the ball can be trapped at the top or bottom edge of the
canvas if it happens to land within the ball's radius of the edge. It
keeps thinking that it's colliding, and therefore reversing the
momentum.

Instead of fixing it immediately, I'd like to use the opportunity to try
demonstrate the issue with a unit test first.