First pass at being able to run mocha tests in the browser.
This is an extremely hacky solution involving opening a page in the
browser to view the results (not CLI), multiple webpack configs and
manually listing all the test files. I also lost the ability to record code coverage somewhere along the way.

It astounds me that it isn't easier to do
TypeScript/Mocha/React/Webpack/Browser without reading dozens of pages
of conflicting documentation. I had a lot of false starts on this.

Future work:
- Combine down to one webpack config
- Avoid having to list all files
- Use karma to provide a CLI for test results