Let the ball bounce from the paddle.
Calculating intersections of objects will be an interesting one to find
a good abstraction for - it seems like a whole-game thing, rather than a
per-object thing. Perhaps each
object needs to be able to return a polygon defining its intersection
area, and an overridable method to define how it behaves during an
overlap event, and a game-level loop that looks for all the possible
object intersections?